Adding toggle for Config Overwrite
Initially we were thinking we would split the module into two modules as per issue #38 .

However, to minimize code duplication we will allow multiple instances of the module. There is now a toggle on the UI for whether the system will overwrite the scan engine settings.

The reccomended usage is to override the scan engine settings on the module using scan engine variables as per the Tip Strip.